requirements.txt
dask
geopandas
shapely
odc-stac
rasterio
xarray
pytest
rtree
matplotlib
rio-stac
stackstac
pystac_client